chore: extract axe-results-to-unified-rules
