+++ src/electron/platform/android/scan-results-to-unified-rules.ts
@@ -3,8 +3,8 @@
 import { UnifiedRule } from 'common/types/store-data/unified-data-interface';
 import { UUIDGenerator } from 'common/uid-generator';
 import { convertAtfsScanResultsToUnifiedRules } from 'electron/platform/android/atfa-results-to-unified-rules';
+import { convertAxeScanResultsToUnifiedRules } from 'electron/platform/android/axe-results-to-unified-rules';
 import { AndroidScanResults } from './android-scan-results';
-import { RuleInformation } from './rule-information';
 import { RuleInformationProviderType } from './rule-information-provider-type';
 
 export type ConvertScanResultsToUnifiedRulesDelegate = (
@@ -25,44 +25,4 @@ export function convertScanResultsToUnifiedRules(
     ).concat(
         convertAtfsScanResultsToUnifiedRules(scanResults, ruleInformationProvider, uuidGenerator),
     );
-}
-
-function convertAxeScanResultsToUnifiedRules(
-    scanResults: AndroidScanResults,
-    ruleInformationProvider: RuleInformationProviderType,
-    uuidGenerator: UUIDGenerator,
-): UnifiedRule[] {
-    if (!scanResults) {
-        return [];
-    }
-
-    const unifiedRules: UnifiedRule[] = [];
-    const ruleIds: Set<string> = new Set();
-
-    for (const result of scanResults.ruleResults) {
-        const ruleId = result.ruleId;
-        if (!ruleIds.has(ruleId)) {
-            const ruleInformation = ruleInformationProvider.getRuleInformation(ruleId);
-
-            if (ruleInformation) {
-                unifiedRules.push(createUnifiedRuleFromRuleResult(ruleInformation, uuidGenerator));
-                ruleIds.add(ruleId);
-            }
-        }
-    }
-
-    return unifiedRules;
-}
-
-function createUnifiedRuleFromRuleResult(
-    ruleInformation: RuleInformation,
-    uuidGenerator: UUIDGenerator,
-): UnifiedRule {
-    return {
-        uid: uuidGenerator(),
-        id: ruleInformation.ruleId,
-        description: ruleInformation.ruleDescription,
-        url: ruleInformation.ruleLink,
-        guidance: ruleInformation.guidance,
-    };
 }
